correction of formula
minor corrections from volts to deci volts in the formula.

also changed the data type of daq_BatteryV_t from float to uint32_t

+++ cmn/daq_common.h
@@ -8,7 +8,7 @@ typedef int32_t daq_DamperPos_t;
 typedef int32_t daq_GearPos_t;
 typedef int32_t daq_SteeringWhlPos_t;
 typedef int32_t daq_Strain_t;
-typedef float daq_BatteryV_t;
+typedef uint32_t daq_BatteryV_t;
 typedef int32_t daq_ThrottlePos_t;
 typedef int32_t daq_FuelPressure_t;
